+++ laps.php
@@ -16,6 +16,10 @@ if ( file_exists( __DIR__ . '/vendor/autoload.php' ) ) {
 	require __DIR__ . '/vendor/autoload.php';
 }
 
+if ( ! class_exists( '\Rarst\Laps\Plugin' ) ) {
+	trigger_error( 'Laps not fully installed! Please install with Composer or download full release archive.', E_USER_ERROR );
+}
+
 $laps = new \Rarst\Laps\Plugin();
 
 $laps->run();